wddm/graphics: new command submission working for 2D, more testing needed
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@50913 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/HostServices/SharedOpenGL/crserverlib/server_presenter.cpp
@@ -371,7 +371,7 @@ static void crFbBltImgScaledRects(const CR_BLITTER_IMG *pSrc, const RTPOINT *pSr
                         pCopyRect->xRight - pCopyRect->xLeft,
                         pCopyRect->yBottom - pCopyRect->yTop,
                         pu8Src,
-                        fSrcInvert ? -pSrc->pitch : pSrc->pitch,
+                        fSrcInvert ? -((GLint)pSrc->pitch) : pSrc->pitch,
                         UnscaledSrcWidth,
                         UnscaledSrcHeight
                         );
